refactor(review): extract auto-commit helpers into Hive::Stages::AutoCommit
Pull the per-pass auto-commit primitives — scope-check, sign-policy,
git-commit invocation — out of `Hive::Stages::Review` into a new
`Hive::Stages::AutoCommit` module so the upcoming stage-exit
`CleanExit` invariant can share one implementation.

This is a pure extract: the existing `fix(review): apply pass NN
findings` commit message + Hive-Fix-* trailers and the
`auto_commit_fix_worktree` flow are byte-identical. Review's
in-file consumers (`auto_commit_*`, `staged_auto_commit_paths`,
`git_head`, `git_command_message`) now delegate to AutoCommit
module-functions, and the legacy `Review::AUTO_COMMIT_*` constants
remain as aliases for any external readers.
